im lost
If Anybody ever reads this kind of comment, i hope you are at least laughing hard while i am suffering, and keep getting it worse for each edit for myself. But since i feel i need to refresh the solutions i tried and save them for later comparison porpuse i think it is neccessary to keep do this with myself.