finish
I give up the frame part. I tried every solution a was able to find, and i even asked gpt to review my code for me. It said, it does not see the problem in neighter index.html, style.css or script.js. I even asked questions about test methods to figure out, but I think i miss something from the beginning of this project and this cause the reason that frame for actual photo does not appear.